+++ backend/src/hardware/cellular/backends/mod.rs
@@ -7,6 +7,7 @@ use std::{error::Error, fmt};
 
 pub mod bearer;
 pub mod config;
+pub mod discovery;
 pub mod io;
 pub mod management;
 pub mod messages;